releasetools: Handle filenames with whitespaces in them
If filenames have whitespces, the default separator (i.e. ' ')
used by make_ext4fs in its block-map is a poor choice.
This patch, in combination with 12ac5cf6aacf19d92d6882fc090297fe01e164ec,
will help in such cases (we now use ':' as the separator between
filenames and their block-lists).

If ':' is not a separator (i.e. split reurns a list of length 1),
we fall back to considering ' ' as the separator.

Change-Id: I0af643fa1bae4f61787996cc451c8f358a38641e